created insert_header function that can be called on each subpage in a new file called global_page_elements
