+++ global_page_elements.py
@@ -1,0 +1,51 @@
+import streamlit as st
+
+# ---- Remove fullscreen buttons from images ----
+def hide_image_fullscreen(): 
+    hide_img_fs = ''' 
+        <style>
+        button[title="View fullscreen"]{
+            visibility: hidden;}
+        </style>
+        '''
+    st.markdown(hide_img_fs, unsafe_allow_html=True)
+
+# ---- LANGUAGE OPTIONS ----
+languages = {"Deutsch": "de", "English": "en"}
+
+def get_language():
+    query_parameters = st.query_params.to_dict()
+    if "lang" not in query_parameters:
+        st.query_params.lang="de"
+        st.experimental_rerun()
+
+def set_language() -> None:
+    if "selected_language" in st.session_state:
+        st.query_params.lang=languages.get(st.session_state["selected_language"]
+        )
+
+def langwrite(english_text, german_text):
+    if st.query_params.lang == 'en':
+        return english_text
+    elif st.query_params.lang == 'de':
+        return german_text
+    return None
+
+# ---- Insert Header Row with Logo and Language Options
+def insert_header():
+    st.markdown(" <style> div[class^='block-container'] { padding-top: 0rem; } </style> ", unsafe_allow_html=True) #removes space at the top of the page
+    get_language()
+    with st.container():
+        left_column, middle_column, right_column = st.columns((4,4,2))
+        with left_column:
+            st.image("ecowhen_logo-name.svg", width=320)
+        with middle_column:
+            st.empty()
+        with right_column:
+            sel_lang = st.radio(
+            "Sprache/ Language",
+            options=languages,
+            horizontal=True,
+            on_change=set_language,
+            key="selected_language",
+    )
